feat: create entities, relationships and layered classes

+++ src/main/java/org/danielmesquita/entities/Category.java
@@ -2,9 +2,8 @@ package org.danielmesquita.entities;
 
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
-import lombok.*;
-
 import java.util.List;
+import lombok.*;
 
 @Entity
 @Table(name = "category")
@@ -15,15 +14,15 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "category_id")
-    private Integer id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "category_id")
+  private Long id;
 
-    @Column(name = "category_name", nullable = false)
-    private String categoryName;
+  @Column(name = "category_name", nullable = false)
+  private String categoryName;
 
-    @OneToMany(mappedBy = "category", fetch = FetchType.LAZY)
-    @JsonManagedReference
-    private List<Product> products;
+  @OneToMany(mappedBy = "category", fetch = FetchType.LAZY)
+  @JsonManagedReference
+  private List<Product> products;
 }
